grapher/requirements.txt
grpcio
grpcio-tools
protobuf
matplotlib==3.10.1
google
google-cloud